+++ timemachine/src/families/turso.rs
@@ -7,6 +7,7 @@ use sql_ast_benchmark::{Parser, ParserId};
 
 macro_rules! turso_version {
     ($name:ident, $cr:ident, $ver:literal, $released:literal) => {
+        #[doc = concat!("turso_parser ", $ver, " release, benchmarked in the time machine.")]
         pub struct $name;
 
         impl Parser for $name {